(PC-37106)[PRO] chore: move indiv. offer summary details screen in appropriate folder (new arch)
