+++ pro/src/components/IndividualOffer/SummaryScreen/SummaryScreen.tsx
@@ -13,6 +13,7 @@ import { getIndividualOfferUrl } from 'commons/core/Offers/utils/getIndividualOf
 import { useNotification } from 'commons/hooks/useNotification'
 import { useOfferWizardMode } from 'commons/hooks/useOfferWizardMode'
 import { getOfferConditionalFields } from 'commons/utils/getOfferConditionalFields'
+import { DisplayOfferInAppLink } from 'components/DisplayOfferInAppLink/DisplayOfferInAppLink'
 import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferNavigation/constants'
 import { OfferAppPreview } from 'components/OfferAppPreview/OfferAppPreview'
 import { RedirectToBankAccountDialog } from 'components/RedirectToBankAccountDialog/RedirectToBankAccountDialog'
@@ -28,7 +29,6 @@ import { SvgIcon } from 'ui-kit/SvgIcon/SvgIcon'
 import { serializeDateTimeToUTCFromLocalDepartment } from '../StocksEventEdition/serializers'
 import { getDepartmentCode } from '../utils/getDepartmentCode'
 
-import { DisplayOfferInAppLink } from './DisplayOfferInAppLink/DisplayOfferInAppLink'
 import { EventPublicationForm } from './EventPublicationForm/EventPublicationForm'
 import { EventPublicationFormValues } from './EventPublicationForm/types'
 import { validationSchema } from './EventPublicationForm/validationSchema'
