+++ pro/src/pages/IndividualOfferSummary/IndividualOfferSummaryDetails/components/IndividualOfferSummaryDetailsScreen.module.scss
