+++ pro/src/pages/IndividualOfferSummary/commons/serializer.ts
